reorganize spec dir
