+++ spec/support/dummy_adapter.rb
